🐛 fix: 빌드 오류 수정
